% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_snv_plots.R
\name{individual_snv_plots}
\alias{individual_snv_plots}
\title{Individual SNV Plots}
\usage{
individual_snv_plots(
  seurat_object,
  processed_snv,
  output_dir = NULL,
  slingshot = TRUE,
  dimensionality_reduction = "UMAP",
  dynamic_cell_size = FALSE,
  save_each_plot = FALSE
)
}
\arguments{
\item{seurat_object}{Processed Seurat object.}

\item{processed_snv}{Data frame of processed SNV information, typically output from the preprocess_snv_data function.}

\item{output_dir}{Directory to save plots and HTML (if save_each_plot is TRUE).}

\item{slingshot}{Logical; whether to include slingshot trajectories. Default: TRUE.}

\item{dimensionality_reduction}{Dimensionality reduction method ('UMAP', 'PCA', 'tSNE'). Default: "UMAP".}

\item{dynamic_cell_size}{Logical; whether to scale cell size dynamically based on SNV and reference read counts. Default: FALSE.}

\item{save_each_plot}{Logical; whether to save each plot individually. Default: FALSE.}
}
\value{
A list containing JSON content for VAF, N_VAR, and N_REF plots.
}
\description{
This function generates individual SNV plots for VAF, N_VAR, and N_REF. It uses processed SNV data
and a Seurat object to create plots with options for saving and including slingshot trajectories.
}
\details{
This function generates individual SNV plots using processed SNV data (processed_snv) and the dimensional
reduction embeddings from a Seurat object. The plots visualize key metrics such as VAF (Variant Allele Fraction),
N_VAR (number of variant reads), and N_REF (number of reference reads).

The plots can be saved individually in the specified output_dir if save_each_plot is set to TRUE.
}
\examples{
# Example usage:
snv_plots <- individual_snv_plots(
  seurat_object = processed_data$SeuratObject,
  processed_snv = processed_data$ProcessedSNV,
  output_dir = "output/individual_plots",
  slingshot = TRUE,
  dimensionality_reduction = "UMAP",
  dynamic_cell_size = FALSE,
  save_each_plot = TRUE
)
}
