% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_snv_data.R
\name{preprocess_snv_data}
\alias{preprocess_snv_data}
\title{Preprocess scSNViz Data}
\usage{
preprocess_snv_data(
  rds_file = NULL,
  countsmatrix_file = NULL,
  snv_file,
  dimensionality_reduction = "UMAP",
  th_vars = 0,
  th_reads = 0,
  enable_sctype = FALSE,
  tissue_type = NULL,
  color_scale = "YlOrRd"
)
}
\arguments{
\item{rds_file}{Path to the RDS file containing a Seurat object (required if countsmatrix_file is not provided).}

\item{countsmatrix_file}{Path to the folder containing STARsolo output files (barcodes.tsv.gz, features.tsv.gz, matrix.mtx.gz).}

\item{snv_file}{Path to the SNV file (required).}

\item{dimensionality_reduction}{The dimensionality reduction method (umap, pca, or tsne). Default: umap.}

\item{th_vars}{Threshold for the number of SNVs per cell. Default: 0.}

\item{th_reads}{Threshold for the number of variant reads per locus. Default: 0.}

\item{enable_sctype}{Logical; enable scType analysis. Default: FALSE.}

\item{tissue_type}{Tissue type for scType. Required if enable_sctype is TRUE.}

\item{color_scale}{Color scale for plots. Options include Blues, Reds, etc. Default: YlOrRd.}
}
\value{
A Seurat object with processed metadata and embeddings.
}
\description{
This function preprocesses SNV and Seurat object data, performs dimensionality reduction,
clustering, and computes various statistics for visualization.
}
\details{
This function reads and processes data from an RDS file or STARsolo counts matrix, along with an SNV file.
\itemize{
\item \strong{RDS File}: A Seurat object is read and processed if provided.
\item \strong{STARsolo Counts Matrix}: If provided, it processes raw counts into a Seurat object.
\item \strong{Dimensionality Reduction}: Options include UMAP (umap), PCA (pca), or t-SNE (tsne).
\item \strong{Thresholding}: Filters cells based on th_vars (number of SNVs) and th_reads (variant reads per locus).
\item \strong{scType Analysis}: If enable_sctype is TRUE, cell-type annotation is performed using the specified tissue_type.
}
}
\examples{
# Example usage:
processed_data <- preprocess_snv_data(
  rds_file = "path/to/seurat.rds",
  snv_file = "path/to/snv_file.tsv",
  dimensionality_reduction = "umap",
  th_vars = 1,
  th_reads = 2,
  enable_sctype = TRUE,
  tissue_type = "Immune system"
)

# Accessing the results:
seurat_object <- processed_data$SeuratObject
plot_data <- processed_data$PlotData
processed_snv <- processed_data$ProcessedSNV

}
